completed material compatible to work with batches.

+++ src/shaders/global/vertex.shader
@@ -8,7 +8,7 @@ layout(location = 3) in float aMaterialIndex;
 centroid out vec2 v_TexCoord;
 centroid out vec3 v_Normal;
 centroid out vec3 v_FragPos;
-centroid out int v_MaterialIndex;
+centroid out float v_MaterialIndex;
 
 uniform mat3 u_NormalMatrix;
 uniform mat4 u_Model;
@@ -16,7 +16,7 @@ uniform mat4 u_View;
 uniform mat4 u_Projection;
 
 void main() {
-	v_MaterialIndex = int(aMaterialIndex);
+	v_MaterialIndex = aMaterialIndex;
 	v_TexCoord = aTexCoord;
 	v_Normal = normalize(u_NormalMatrix * aNormal);
 	v_FragPos = vec3(u_Model * vec4(aPosition, 1.0));
